Renamed ReactiveClient -> RxClient, FutureClient -> AsyncClient #149
