+++ src/main/java/org/mqttbee/mqtt/mqtt3/Mqtt3ClientView.java
@@ -24,7 +24,7 @@ import io.reactivex.functions.Function;
 import org.jetbrains.annotations.NotNull;
 import org.mqttbee.api.mqtt.MqttGlobalPublishFlowType;
 import org.mqttbee.api.mqtt.mqtt3.Mqtt3ClientData;
-import org.mqttbee.api.mqtt.mqtt3.Mqtt3ReactiveClient;
+import org.mqttbee.api.mqtt.mqtt3.Mqtt3RxClient;
 import org.mqttbee.api.mqtt.mqtt3.message.connect.Mqtt3Connect;
 import org.mqttbee.api.mqtt.mqtt3.message.connect.connack.Mqtt3ConnAck;
 import org.mqttbee.api.mqtt.mqtt3.message.publish.Mqtt3Publish;
@@ -53,7 +53,7 @@ import org.mqttbee.util.MustNotBeImplementedUtil;
  * @author Silvio Giebl
  * @author David Katz
  */
-public class Mqtt3ClientView implements Mqtt3ReactiveClient {
+public class Mqtt3ClientView implements Mqtt3RxClient {
 
     private static final @NotNull Function<Throwable, Completable> EXCEPTION_MAPPER_COMPLETABLE =
             e -> Completable.error(Mqtt3ExceptionFactory.map(e));
